First pass at my own X-ends
Trying to make a prusa x end that doesn't rely on layer adhesion so
much.

Using cheap and not-very-good zeni kinetic filament (I would not
recommend it) resulted in an x-end that delaminated.

Parts that I printed using ProtoParadigm did not have this problem, but
it was probable being strained in a not-so-good way.